- Fix for previous UUID commit (missing uint128 cleanup). - Cleaned up and updated some vector types (sint32v2, sint32v4, sint16v2, etc.)

+++ src/core/types/float4.cpp
@@ -85,7 +85,7 @@ float3 float4::xyz(void) const
    
 void float4::normalize(void)
 {
-    float tmp = sqrt(x*x + y*y + z*z);
+    float tmp = sqrtf(x*x + y*y + z*z);
     if (tmp) 
     {			
         x /= tmp;
